backend/requirements.txt
# FastAPI Backend Dependencies
# Install with: pip install -r requirements.txt

# ============================================================================
# FastAPI Core Dependencies
# ============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# ============================================================================
# Async and Scheduling
# ============================================================================
APScheduler>=3.10.4
asyncio>=3.4.3

# ============================================================================
# Docker Integration
# ============================================================================
docker>=7.0.0

# ============================================================================
# Server-Sent Events (SSE)
# ============================================================================
sse-starlette>=1.8.2

# ============================================================================
# Existing Scraper Dependencies
# ============================================================================
# Web scraping
playwright>=1.40.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1

# Database
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# ============================================================================
# Additional Utilities
# ============================================================================
# HTTP client (included with supabase but explicit for clarity)
httpx>=0.26.0

# Logging
python-json-logger>=2.0.7

# Date parsing
python-dateutil>=2.8.2

# ============================================================================
# Development Dependencies (Optional)
# ============================================================================
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.12.0
# flake8>=7.0.0
# mypy>=1.8.0

